+++ plugins/org.eclipse.actf.model.ui.editors.ie/src/org/eclipse/actf/model/internal/ui/editors/ie/events/impl/BeforeNavigate2ParametersImpl.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2007 IBM Corporation and Others
+ * Copyright (c) 2007, 2019 IBM Corporation and Others
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -29,7 +29,7 @@ public class BeforeNavigate2ParametersImpl extends AbstractEventParameters imple
     public BeforeNavigate2ParametersImpl(OleEvent event) {
         super(event);
         if( Platform.inDebugMode() ) {
-            int     dispBrowser =     getBrowserAddress();
+            long     dispBrowser =     getBrowserAddress();
             String  url =             getUrl();
             int     flags =           getFlags();
             String  targetFrameName = getTargetFrameName();
@@ -43,7 +43,7 @@ public class BeforeNavigate2ParametersImpl extends AbstractEventParameters imple
     /* (non-Javadoc)
      * @see org.eclipse.actf.model.ui.editors.ie.events.BeforeNavigate2Parameters#getBrowserAddress()
      */
-    public int getBrowserAddress() {
+    public long getBrowserAddress() {
         return getDispatchAddress(INDEX_Browser);
     }
