+++ plugins/org.eclipse.actf.model.ui.editors.ie/src/org/eclipse/actf/model/internal/ui/editors/ie/events/impl/DocumentCompleteParametersImpl.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2007 IBM Corporation and Others
+ * Copyright (c) 2007, 2019 IBM Corporation and Others
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -27,7 +27,7 @@ public class DocumentCompleteParametersImpl extends AbstractEventParameters impl
     public DocumentCompleteParametersImpl(OleEvent event) {
         super(event);
         if( Platform.inDebugMode() ) {
-            int     dispWindow = getWindowAddress();
+            long     dispWindow = getWindowAddress();
             String  url =        getUrl();
             boolean isTop =      isTopWindow();
             System.out.println("DocumentComplete("+dispWindow+",\""+url+"\") top="+isTop); //$NON-NLS-1$ //$NON-NLS-2$ //$NON-NLS-3$
@@ -37,7 +37,7 @@ public class DocumentCompleteParametersImpl extends AbstractEventParameters impl
     /* (non-Javadoc)
      * @see org.eclipse.actf.model.ui.editors.ie.events.DocumentCompleteParameters#getWindowAddress()
      */
-    public int getWindowAddress() {
+    public long getWindowAddress() {
         return getDispatchAddress(INDEX_Window);
     }
