+++ plugins/org.eclipse.actf.model.ui.editors.ie/src/org/eclipse/actf/model/internal/ui/editors/ie/events/impl/NavigateErrorParametersImpl.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2007 IBM Corporation and Others
+ * Copyright (c) 2007, 2019 IBM Corporation and Others
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -30,7 +30,7 @@ public class NavigateErrorParametersImpl extends AbstractEventParameters impleme
     public NavigateErrorParametersImpl(OleEvent event) {
         super(event);
         if( Platform.inDebugMode() ) {
-            int     dispBrowser =     getBrowserAddress();
+            long     dispBrowser =     getBrowserAddress();
             String  url =             getUrl();
             String  targetFrameName = getTargetFrameName();
             int     statusCode =      getStatusCode();
@@ -42,7 +42,7 @@ public class NavigateErrorParametersImpl extends AbstractEventParameters impleme
     /* (non-Javadoc)
      * @see org.eclipse.actf.model.ui.editors.ie.events.NavigateErrorParameters#getBrowserAddress()
      */
-    public int getBrowserAddress() {
+    public long getBrowserAddress() {
         return getDispatchAddress(INDEX_Browser);
     }
