+++ plugins/org.eclipse.actf.util.win32/src/org/eclipse/actf/util/win32/VariantUtil.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2008 IBM Corporation and Others
+ * Copyright (c) 2008, 2019 IBM Corporation and Others
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -22,7 +22,7 @@ public class VariantUtil {
 	 *            address of IDispatch
 	 * @return Variant object which represents an IDispatch interface
 	 */
-	public static Variant createVariantFromIDispatchAddress(int address) {
+	public static Variant createVariantFromIDispatchAddress(long address) {
 		return new Variant(new IDispatch(address));
 	}
